Refactor google settings to reusable parts
